+++ core/types/audio.ts
@@ -1,6 +1,6 @@
 /**
  * File: core/types/audio.ts
- * Version: 1.7.32
+ * Version: 1.7.37
  * Author: Aura Vision Team
  * Copyright (c) 2024 Aura Vision. All rights reserved.
  * Updated: 2025-03-05 12:00